improved video making process

+++ src/VideoMaker.py
@@ -8,30 +8,36 @@ from src.TwitchAuthenticator import TwitchAuthenticator
 from src.AudioToText import AudioToText
 import datetime
 import json
+import asyncio
 
-TRANSITION_VOLUME = 0.1
-TRANSITION_SUBCLIP = (2, 2.5)
-OUTPUT_RESOLUTION = (1920, 1080)
-SORT_BY_TIME = True
-DEFAULT_TRANSITION_DIR = 'video_assets/transition.mp4'
-DEFAULT_OUTPUT_DIR = 'temp/output/'
-DEFAULT_CLIP_DIR = 'temp/clips/'
-DEFAULT_CLIP_INFO_DIR = 'clip_info/'
-DEFAULT_READY_CHANNEL_CSV = 'clip_info/saturated_channels.csv'
-TARGET_VIDEO_LENGTH = 1 * 60 # 1 minutes
-LOOK_BACK_TIME = 24 * 7 # 1 week
-REQUIRED_CLIP_NUM = 10
+with open('config.json') as f:
+    config = json.load(f)
 
-# these do not need to be changed
-SPOKEN_WORD_TRIM = 3.0
+TRANSITION_VOLUME = config['TRANSITION_VOLUME']
+TRANSITION_SUBCLIP = tuple(config['TRANSITION_SUBCLIP'])
+OUTPUT_RESOLUTION = tuple(config['OUTPUT_RESOLUTION'])
+SORT_BY_TIME = config['SORT_BY_TIME']
+DEFAULT_TRANSITION_DIR = config['DEFAULT_TRANSITION_DIR']
+DEFAULT_OUTPUT_DIR = config['DEFAULT_OUTPUT_DIR']
+DEFAULT_CLIP_DIR = config['DEFAULT_CLIP_DIR']
+DEFAULT_CLIP_INFO_DIR = config['DEFAULT_CLIP_INFO_DIR']
+DEFAULT_READY_CHANNEL_CSV = config['DEFAULT_READY_CHANNEL_CSV']
+TARGET_VIDEO_LENGTH = config['TARGET_VIDEO_LENGTH']
+LOOK_BACK_TIME = config['LOOK_BACK_TIME']
+REQUIRED_CLIP_NUM = config['REQUIRED_CLIP_NUM']
+TRANSITION_THRESHOLD = config['TRANSITION_THRESHOLD']
+
+# these are likly not going to change so they are not in the config file
+EDGE_TRIM_THRESHOLD = 3.0
+MIDDLE_TRIM_THRESHOLD = 5.0
 MIN_CLIP_LENGTH = 5.0
 
 
 class VideoMaker:
-    def __init__(self, clip_dirs, output_dir, intro_clip_dir=None, outro_clip_dir=None, transition_dir=None):
+    def __init__(self, clips, output_dir, intro_clip_dir=None, outro_clip_dir=None, transition_dir=None):
         self.intro_clip_dir = intro_clip_dir
         self.outro_clip_dir = outro_clip_dir
-        self.clip_dirs = clip_dirs
+        self.clips = clips
         self.output_dir = output_dir
         self.transition_dir = transition_dir
 
@@ -39,15 +45,15 @@ class VideoMaker:
 
     # makes a video from the clips Returns True if successful
     def make_video(self):
-        # first thing lets combine all the clips in clip_dirs
+        # second thing lets combine all the clips in clip_dirs
         videos = []
-        for i, clip_dir in enumerate(self.clip_dirs):
-            if not os.path.exists(clip_dir):
-                print("Clip " + clip_dir + " does not exist")
+        for i, clip in enumerate(self.clips):
+            if not os.path.exists(clip.clip_dir):
+                print("Clip " + clip.clip_dir + " does not exist")
                 continue
 
             # run the video through the filter to get the subclip we want
-            filter_result = self.filter_clips(clip_dir)
+            filter_result = self.filter_clips(clip.clip_dir)
 
             # if None the clip was completely rejected
             if filter_result == None:
@@ -56,11 +62,15 @@ class VideoMaker:
             # add the clip to the list of clips
             videos.append(filter_result)
 
-            videos.append(VideoFileClip(self.transition_dir).volumex(TRANSITION_VOLUME).subclip(TRANSITION_SUBCLIP[0], TRANSITION_SUBCLIP[1]))
+            # find the time between this clip and the next clip
+            time_between_clips = datetime.timedelta(seconds=TRANSITION_THRESHOLD)
+            if (i != len(self.clips) - 1):
+                time_between_clips = self.clips[i + 1].time - self.clips[i].time - datetime.timedelta(seconds=videos[-1].duration)
 
-        # if this is the last clip then we don't the transition
-        if (len(videos) > 0):
-            videos.pop()    
+            # add transitions between clips with a time greater than 
+            if (i != len(self.clips) - 1 or time_between_clips >= datetime.timedelta(seconds=TRANSITION_THRESHOLD)):
+                videos.append(VideoFileClip(self.transition_dir).volumex(TRANSITION_VOLUME).subclip(TRANSITION_SUBCLIP[0], TRANSITION_SUBCLIP[1]))
+
 
         # now lets add the intro and outro clips
         if (self.intro_clip_dir != None):
@@ -85,7 +95,7 @@ class VideoMaker:
         # TODO
 
         # get the streamer name from the first clip
-        streamer_name = self.clip_dirs[0].split('/')[-1].split('_')[0]
+        streamer_name = self.clips[0].clip_dir.split('/')[-1].split('_')[0]
 
         # render the video
         final_clip.write_videofile(self.output_dir + streamer_name + datetime.datetime.now().strftime("%Y-%m-%d-%H-%M-%S") + '.mp4')
@@ -112,19 +122,31 @@ class VideoMaker:
 
         print("Clip length before trim: " + str(clip_video.duration))
 
+        # trim the clip to remove silence at the beginning and end
         clip_video = self.trim_clip(clip_video, text_times)
 
         # if the clip was rejected then return None
         if clip_video == None:
             print("Clip to short after trimming rejecting...")
             return None
-        
         print("Clip length after trim: " + str(clip_video.duration))
+
+        # remove extra silence in the middle of the clip
+        clip_video = self.trim_silence(clip_video, text_times)
+
+        # if the clip was rejected then return None
+        if clip_video == None:
+            print("Clip to short after trimming rejecting...")
+            return None
+        print("Clip length after middle trim: " + str(clip_video.duration))
 
         return clip_video
     
     # trims the clip to remove silence at the end and beginning
     def trim_clip(self, clip, text_times):
+        if clip == None:
+            return None
+
         # the variable to store the desired start and end times
         start_time = 0
         end_time = 0
@@ -132,13 +154,13 @@ class VideoMaker:
         # filter out the begining and end silence
         for i in range(len(text_times)):
             # if a word took unusally long to say we know to filter that section out
-            if (text_times[i]["end"] - text_times[i]["start"] < SPOKEN_WORD_TRIM):
+            if (text_times[i]["end"] - text_times[i]["start"] < EDGE_TRIM_THRESHOLD):
                 start_time = text_times[i]["start"]
                 break
         
         for i in range(len(text_times) - 1, 0, -1):
             # if a word took unusally long to say we know to filter that section out
-            if (text_times[i]["end"] - text_times[i]["start"] < SPOKEN_WORD_TRIM):
+            if (text_times[i]["end"] - text_times[i]["start"] < EDGE_TRIM_THRESHOLD):
                 end_time = text_times[i]["end"]
                 break
 
@@ -149,11 +171,40 @@ class VideoMaker:
         
         return clip.subclip(start_time, end_time)
 
+    # trims the clip to remove silence in the middle
+    def trim_silence(self, clip, text_times):
+        if clip == None:
+            return None
+        
+        # loop over the text and cut the words that take longer than the threshold
+        clips = []
+        last_end_time = 0
+        for i in range(len(text_times)):
+            # if a word took unusally long to say we know to filter that section out
+            if (text_times[i]["end"] - text_times[i]["start"] > MIDDLE_TRIM_THRESHOLD):
+                clips.append(clip.subclip(last_end_time, text_times[i]["start"] + 0.1))
+                last_end_time = self.clamp(text_times[i]["end"], 0, clip.duration)
 
+        # append the last clip
+        if (clip.duration - last_end_time > MIDDLE_TRIM_THRESHOLD):
+            clips.append(clip.subclip(last_end_time, clip.duration))
+
+        # if there are no clips return None
+        if (len(clips) == 0):
+            return None
+        
+        # now we need to combine the clips
+        final_clip = concatenate_videoclips(clips, method="compose")
+
+        return final_clip
+    
+    def clamp(self, n, minn, maxn):
+        return max(min(maxn, n), minn)
+                
     # makes a video from a directory of clips and uses the default transition and content for the channel 
     # specified by channel_name
     @staticmethod
-    async def make_from_channel(channel_name, clip_count=15, output_dir=DEFAULT_OUTPUT_DIR, transition_dir=DEFAULT_TRANSITION_DIR):
+    async def make_from_channel(channel_name, clip_count=10, output_dir=DEFAULT_OUTPUT_DIR, transition_dir=DEFAULT_TRANSITION_DIR, time=LOOK_BACK_TIME, sort_by_time=SORT_BY_TIME):
         # init the twitch authenticator
         authenticator = TwitchAuthenticator()
 
@@ -171,10 +222,14 @@ class VideoMaker:
 
         print("Getting clips for " + users[0].display_name)
 
-        clips = clipGetter.get_clips(users[0], authenticator.get_client(), time=LOOK_BACK_TIME, clip_dir=DEFAULT_CLIP_DIR, clip_count=clip_count, sort_by_time=SORT_BY_TIME)
-        dirs = [clip.clip_dir for clip in clips]
+        clips = clipGetter.get_clips(users[0], authenticator.get_client(), time=time, clip_dir=DEFAULT_CLIP_DIR, clip_count=clip_count, sort_by_time=sort_by_time)
 
         print("Got clips... making video")
+
+        # check for overlaping clips
+        clip_compiler = ClipCompiler()
+        clips = clip_compiler.merge_clips(clips)
+        dirs = [clip.clip_dir for clip in clips]
 
         # TODO cross reference the clipinfo files to see if there are any of our clips in the top num 
 
@@ -183,17 +238,21 @@ class VideoMaker:
             os.makedirs(output_dir)
 
         # now lets make the video
-        video_maker = VideoMaker(dirs, output_dir, transition_dir=transition_dir)
+        video_maker = VideoMaker(clips, output_dir, transition_dir=transition_dir)
 
         status = video_maker.make_video()
 
         if not status:
             return False
+        
+        # sleep for a bit so we don't have issues deleting the clips
+        await asyncio.sleep(5)
 
         # delete the clips
         for dir_ in dirs:
             try:
-                os.remove(dir_)
+                if os.path.exists(dir_):
+                    os.remove(dir_)
             except OSError:
                 # if one file is not deleted its not a big deal
                 pass
@@ -201,6 +260,10 @@ class VideoMaker:
     # makes a video from a csv file that specifies the clips to use csv_dir can be a csv file or a directory of csv files
     @staticmethod
     async def make_from_csv(csv_dir, output_dir=DEFAULT_OUTPUT_DIR , transition_dir=DEFAULT_TRANSITION_DIR):
+        # combine any overlaping clips
+        clip_compiler = ClipCompiler()
+        clip_compiler.merge_clips(csv_dir)
+        
         # load in the csv
         clips = []
         with open(csv_dir, 'r') as f:
@@ -218,16 +281,14 @@ class VideoMaker:
             clips.sort(key=lambda x: x.time)
 
         # get the clip dirs
-        dirs = []
-        for clip in clips:
-            dirs.append(clip.clip_dir)
+        dirs = [clip.clip_dir for clip in clips]
 
         # make sure the output dir exists
         if not os.path.exists(output_dir):
             os.makedirs(output_dir)
         
         # now lets make the video
-        video_maker = VideoMaker(dirs, output_dir, transition_dir=transition_dir)
+        video_maker = VideoMaker(clips, output_dir, transition_dir=transition_dir)
 
         status = video_maker.make_video()
 
@@ -237,7 +298,8 @@ class VideoMaker:
         # delete the clips
         for dir_ in dirs:
             try:
-                os.remove(dir_)
+                if os.path.exists(dir_):
+                    os.remove(dir_)
             except OSError:
                 # if one file is not deleted its not a big deal
                 pass
@@ -251,26 +313,12 @@ class VideoMaker:
     async def make_from_csvs(clip_info=DEFAULT_CLIP_INFO_DIR):
         # get a count of the videos per channel by reading the csv's in the clip info dir
         dirs = os.listdir(clip_info)
-        
-        for dir_ in dirs:
-            print(dir_)
-
-        return
-
-        # combine any clips that overlap
-        clip_compiler = ClipCompiler()
-        for dir_ in dirs:
-            clip_compiler.merge_clips(dir_)
 
         # make a video for each channel
         for dir_ in dirs:
-            await VideoMaker.make_from_csv(dir_)
+            await VideoMaker.make_from_csv(clip_info + '/' + dir_)
 
         # remove all the clips from the csv
-        with open(csv_dir, 'w') as f:
-            f.write('')
-
-
-
-
-
+        for dir_ in dirs:
+            with open(dir_, 'w') as f:
+                f.write('')
